feat: 웹소켓 연결 확인 & user 정보 전달 & 전달 데이터 정리

+++ src/main/java/com/ddhouse/chat/service/ChatService.java
@@ -61,16 +61,31 @@ public class ChatService {
         return chatMessageRepository.findAllByRoomId(roomId);
     }
     @Transactional
-    public void deleteChatRoom(Long roomId){
-        // TODO : 카산드라 db 내용 해당 채팅방 데이터들을 먼저 다 지우기
-        findMessagesByRoomId(roomId)
-                .flatMap(chatMessage -> {
-                    System.out.println(chatMessage.getId());  // 아이디 출력
-                    return chatMessageRepository.deleteById(chatMessage.getId());  // 삭제 작업
-                })
-                .then();
-        userChatRoomRepository.deleteByChatRoomId(roomId);
-        // 해당 채팅방 지우기
-        chatRoomRepository.deleteById(roomId);
+    public void deleteChatRoom(Long roomId, Long myId){
+        // 채팅 이용자가 0명인 경우 전체 대화 삭제
+        Optional<ChatRoom> chatRoomOptional = chatRoomRepository.findById(roomId);
+
+        chatRoomOptional.ifPresent(chatRoom -> {
+            chatRoom.decreaseMemberNum(); // memberNum 감소
+
+            if (chatRoom.getMemberNum() == 0) {
+                // TODO : 카산드라 db 내용 해당 채팅방 데이터들을 먼저 다 지우기
+                findMessagesByRoomId(roomId)
+                        .flatMap(chatMessage -> {
+                            System.out.println(chatMessage.getId());  // 아이디 출력
+                            return chatMessageRepository.deleteById(chatMessage.getId());  // 삭제 작업
+                        })
+                        .then();
+                userChatRoomRepository.deleteByChatRoomId(roomId);
+                // 해당 채팅방 지우기
+                chatRoomRepository.deleteById(roomId);
+            } else {
+                // TODO : 채팅방 나가기 한 사용자에게는 해당 채팅방이 보이지 않아야함.
+            }
+
+            chatRoomRepository.save(chatRoom); // 변경사항 저장
+        });
+
+
     }
 }
